feat(营销中心): 优惠券 1.支持商品关联创建优惠券 2.查询有效优惠券关联商品

+++ src/main/java/com/acme/acmemall/model/FootprintVo.java
@@ -24,7 +24,7 @@ public class FootprintVo implements Serializable {
     //会员Id
     private Long user_id;
     //商品id
-    private Integer goods_id;
+    private Long goods_id;
     //记录时间
     private Long add_time;
     //推荐人
